Add BApp abstraction for go-prompt and promptui
